+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## 0.7.0
+
 - Add support for textDocument/rename feature. #6
 - Add support for workspace/applyEdit. #7